+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased](https://github.com/python-social-auth/social-app-django/commits/master)
 
 ### Changed
+- Remove usage of set/get current strategy methods
 - Remove usage of `social_auth` related name since it should be consider a simple helper.
 
 ## [0.1.0](https://github.com/python-social-auth/social-app-django/releases/tag/0.1.0) - 2016-12-28